@@ -22,6 +22,12 @@ CACHE = Path.home() / ".mlmc" / "datasets"
 URL = "https://aspra29.informatik.uni-leipzig.de:9090/"
 
 def _load_from_tmp(dataset):
+    """
+    Loads a dataset from cache.
+
+    :param dataset: Name of the dataset
+    :return: Tuple of form (data, classes) if dataset exists in cache, else None
+    """
     if not Path.exists(CACHE):
         Path.mkdir(CACHE)
     if Path.is_file(CACHE / dataset):
@@ -33,12 +39,19 @@ def _load_from_tmp(dataset):
         return None
 
 def _save_to_tmp(dataset, data):
+    """
+    Saves a dataset to cache.
+
+    :param dataset: Name of the dataset
+    :param data: Tuple of form (data, classes)
+    :return: Path to the saved dataset if dataset didn't exist in cache, else False.
+    """
     if not Path.exists(CACHE):
         Path.mkdir(CACHE)
     if not Path.is_file(CACHE / dataset):
         with open(CACHE / dataset, "wb") as f:
             pickle.dump(data, f)
-        return True
+        return CACHE / dataset
     else:
         return False
 
@@ -48,6 +61,12 @@ def _save_to_tmp(dataset, data):
 
 
 def load_aapd():
+    """
+    Loads AAPD (Arxiv Academic Paper Dataset) from cache. If it doesn't exist in cache the dataset will be downloaded.
+    Task: Multilabel Classification
+
+    :return: Tuple of form (data, classes)
+    """
     data = _load_from_tmp("aapd")
     if data is not None:
         return data
@@ -76,6 +95,14 @@ def load_aapd():
 
 
 def load_rcv1(path=None):
+    """
+    Loads rcv1 (Reuters Corpus Volume I) from cache. This dataset can't be downloaded automatically so a path to the
+    archive has to be provided when first called.
+    Task: Multilabel Classification
+
+    :param path: Path to unprocessed corpus
+    :return: Tuple of form (data, classes)
+    """
     data = _load_from_tmp("rcv1")
     if data is not None:
         return data
@@ -147,7 +174,8 @@ def load_rcv1(path=None):
     import re
     edges = [(re.split(" +", x)[1], re.split(" +", x)[3]) for x in content]
     edges = [(topicmap.get(x[0],x[0]).capitalize(),topicmap.get(x[1],x[1]).capitalize()) for x in edges]
-    graph = nx.DiGraph(edges[1:])
+    graph = nx.OrderedDiGraph(edges[1:])
+    graph.remove_node("Root")
     data["graph"] = graph
 
 
@@ -162,6 +190,14 @@ def load_rcv1(path=None):
 
 
 def load_wiki30k(path="/disk1/users/jborst/Data/Test/MultiLabel/wiki30k"):
+    """
+    Loads wiki30k from cache. This dataset can't be downloaded automatically so a path to the archive has to be provided
+    when first called.
+    Task: Multilabel Classification
+
+    :param path: Path to pickled dataset
+    :return: Tuple of form (data, classes)
+    """
     import pickle
     with open(Path(path) / "wiki30k_raw_text.p", "rb") as f:
         content = pickle.load(f)
@@ -174,6 +210,12 @@ def load_wiki30k(path="/disk1/users/jborst/Data/Test/MultiLabel/wiki30k"):
 
 
 def load_eurlex():
+    """
+    Loads EUR-Lex from cache. If it doesn't exist in cache the dataset will be downloaded.
+    Task: Multilabel Classification
+
+    :return: Tuple of form (data, classes)
+    """
     data = _load_from_tmp("eurlex")
     if data is not None:
         return data
@@ -205,6 +247,13 @@ def load_eurlex():
 
 
 def load_huffpost(test_split=0.25):
+    """
+    Loads Huffington Post dataset from cache. If it doesn't exist in cache the dataset will be downloaded.
+    Task: Multilabel Classification
+
+    :param test_split: Size of test split as fraction of 1
+    :return: Tuple of form (data, classes)
+    """
     data = _load_from_tmp("huffpost")
     if data is not None:
         return data
@@ -236,6 +285,13 @@ def load_huffpost(test_split=0.25):
 
 
 def load_moviesummaries(test_split=0.25):
+    """
+    Loads Movie Summaries dataset from cache. If it doesn't exist in cache the dataset will be downloaded.
+    Task: Multilabel Classification
+
+    :param test_split: Size of test split as fraction of 1
+    :return: Tuple of form (data, classes)
+    """
     data = _load_from_tmp("moviesummaries")
     if data is not None:
         return data
@@ -284,6 +340,12 @@ def load_moviesummaries(test_split=0.25):
 # ----------------------------------------------
 
 def load_blurbgenrecollection():
+    """
+    Loads Blurb Genre Collection dataset from cache. If it doesn't exist in cache the dataset will be downloaded.
+    Task: Multilabel Classification
+
+    :return: Tuple of form (data, classes)
+    """
     url = "https://fiona.uni-hamburg.de/ca89b3cf/blurbgenrecollectionen.zip"
     data = _load_from_tmp("blurbgenrecollection")
     if data is not None:
@@ -309,7 +371,7 @@ def load_blurbgenrecollection():
         classes = list(set([x for y in data["train"][1] + data["valid"][1] + data["test"][1] + edges for x in y]))
         classes = dict(zip(classes, range(len(classes))))
         edges = [e for e in edges if len(e) == 2]
-        graph = nx.DiGraph()
+        graph = nx.OrderedDiGraph()
         graph.add_nodes_from(classes.keys())
         graph.add_edges_from(edges)
         data["graph"] = graph
@@ -318,6 +380,13 @@ def load_blurbgenrecollection():
 
 
 def load_blurbgenrecollection_de():
+    """
+    Loads German version of Blurb Genre Collection dataset from cache. If it doesn't exist in cache the dataset will be
+    downloaded.
+    Task: Multilabel Classification
+
+    :return: Tuple of form (data, classes)
+    """
     url = "https://www.inf.uni-hamburg.de/en/inst/ab/lt/resources/data/germeval-2019-hmc/germeval2019t1-public-data-final.zip"
     data = _load_from_tmp("blurbgenrecollection_de")
     if data is not None:
@@ -346,7 +415,7 @@ def load_blurbgenrecollection_de():
         classes = list(set([x for y in data["train"][1] + data["valid"][1] + data["test"][1] + edges for x in y]))
         classes = dict(zip(classes, range(len(classes))))
         edges = [e for e in edges if len(e) == 2]
-        graph = nx.DiGraph(edges)
+        graph = nx.OrderedDiGraph(edges)
         data["graph"] = graph
         _save_to_tmp("blurbgenrecollection_de", (data, classes))
         return data, classes
@@ -368,6 +437,12 @@ def load_webofscience():
     #     zipfile = ZipFile(BytesIO(resp.read()))
 
 def load_20newsgroup():
+    """
+    Loads 20newsgroup dataset from cache. If it doesn't exist in cache the dataset will be downloaded.
+    Task: Singlelabel Classification
+
+    :return: Tuple of form (data, classes)
+    """
     url = "http://qwone.com/~jason/20Newsgroups/20news-bydate.tar.gz"
     data = _load_from_tmp("20newsgroup")
     if data is not None:
@@ -411,13 +486,19 @@ def load_20newsgroup():
                 for x in classes for y in range(len(x.split(".")) - 1)]
             edges += list(set([("ROOT", x.split(".")[0]) for x in classes]))
 
-            graph = nx.DiGraph(edges)
+            graph = nx.OrderedDiGraph(edges)
             data = {"train": traindata, "test": testdata, "graph": graph}
         _save_to_tmp("20newsgroup", (data, classes))
 
         return data, classes
 
 def load_agnews():
+    """
+    Loads AG News dataset from cache. If it doesn't exist in cache the dataset will be downloaded.
+    Task: Singlelabel Classification
+
+    :return: Tuple of form (data, classes)
+    """
     url = "https://s3.amazonaws.com/fast-ai-nlp/ag_news_csv.tgz"
     data = _load_from_tmp("agnews")
     if data is not None:
@@ -459,6 +540,12 @@ def load_agnews():
         return data, classes
 
 def load_dbpedia():
+    """
+    Loads DBpedia dataset from cache. If it doesn't exist in cache the dataset will be downloaded.
+    Task: Singlelabel Classification
+
+    :return: Tuple of form (data, classes)
+    """
     url = "https://github.com/le-scientifique/torchDatasets/raw/master/dbpedia_csv.tar.gz"
     data = _load_from_tmp("dbpedia")
     if data is not None:
@@ -500,6 +587,12 @@ def load_dbpedia():
         return data, classes
 
 def load_ohsumed():
+    """
+    Loads Ohsumed dataset from cache. If it doesn't exist in cache the dataset will be downloaded.
+    Task: Multilabel Classification
+
+    :return: Tuple of form (data, classes)
+    """
     url = "http://disi.unitn.eu/moschitti/corpora/ohsumed-first-20000-docs.tar.gz"
     url_classes = "http://disi.unitn.eu/moschitti/corpora/First-Level-Categories-of-Cardiovascular-Disease.txt"
     data = _load_from_tmp("ohsumed")
@@ -588,6 +681,13 @@ def load_ohsumed():
         return data, classes
 
 def export(data, classes, path=Path("./export")):
+    """
+    Exports the data and class dictionaries of a dataset to text files.
+
+    :param data: Data dictionary containing training, validation and test splits
+    :param classes: Class dictionary mapping labels to indices
+    :param path: Path to write the text files to
+    """
     path = Path(path)
     if not path.exists():
         path.mkdir()
@@ -604,3 +704,41 @@ def export(data, classes, path=Path("./export")):
 
     with open(path / "classes.txt", "w") as o:
         o.writelines([x + "\n" for x in classes.keys()])
+
+def load_yahoo_answers():
+    """
+    Loads Yahoo Answers dataset from cache. If it doesn't exist in cache the dataset will be downloaded.
+    Task: Singlelabel Classification
+
+    :return: Tuple of form (data, classes)
+    """
+    url = (URL+"/yahoo_answers_csv.tar.gz").replace("https","http")
+    data = _load_from_tmp("yahoo_answers")
+    if data is not None:
+        return data
+    else:
+        import csv
+        with tempfile.TemporaryDirectory() as tmpdir:
+            resp = urlopen(url)
+            tf = tarfile.open(fileobj=resp, mode="r|gz")
+            tf.extractall(Path(tmpdir))
+
+            with open(Path(tmpdir) /'yahoo_answers_csv'/'classes.txt',"r") as f:
+                classes = [x.replace("\n","") for x in f.readlines()]
+                classes_rev =dict(zip(range(len(classes)) ,classes))
+                classes = dict(zip(classes, range(len(classes))))
+
+
+
+            with open(Path(tmpdir) /'yahoo_answers_csv'/'test.csv',"r") as f:
+                test = [(",".join([s for s in x[1:] if s != ""]), [classes_rev[int(x[0]) - 1]]) for x in csv.reader(f, dialect="unix")]
+            with open(Path(tmpdir) /'yahoo_answers_csv'/'train.csv', "r") as f:
+                train = [(",".join([s for s in x[1:] if s != ""]), [classes_rev[int(x[0]) - 1]]) for x in csv.reader(f, dialect="unix")]
+
+
+        data = {
+            "train": ([x[0] for x in train], [x[1] for x in train]),
+            "test": ([x[0] for x in test], [x[1] for x in test]),
+        }
+        _save_to_tmp("yahoo_answers", (data, classes))
+        return data, classes
