@@ -3,54 +3,83 @@ https://raw.githubusercontent.com/EMNLP2019LSAN/LSAN/master/attention/model.py
 """
 import torch
 import torch.nn.functional as F
-from .abstracts import TextClassificationAbstract
-from ..representation import get
+from mlmc.models.abstracts.abstracts import TextClassificationAbstract
+from ..representation import is_transformer
 
-class LSANOriginal(TextClassificationAbstract):
+class LSAN(TextClassificationAbstract):
     """
     https://raw.githubusercontent.com/EMNLP2019LSAN/LSAN/master/attention/model.py
     """
-    def __init__(self, classes, representation, label_embed=None, label_freeze=True, lstm_hid_dim=300, d_a=200 ,max_len=500,**kwargs):
-        super(LSANOriginal, self).__init__(**kwargs)
+    def __init__(self, label_embedding=None, label_freeze=False, lstm_hid_dim=300, d_a=200,  **kwargs):
+        """
+        Class constructor and initialization of every hyperparameter.
+
+        :param label_embedding: If not None, loads the specified label embeddings instead of creating them from scratch.
+        :param label_freeze: If True, the embedding tensor does not get updated in the learning process.
+        :param lstm_hid_dim: Hidden state dimension of the LSTM used to create the word embeddings
+        :param d_a: Arbitrarily set hyperparameter
+        :param kwargs: Optimizer and loss function keyword arguments, see `mlmc.models.abstracts.abstracts.TextClassificationAbstract`
+        """
+        super(LSAN, self).__init__(**kwargs)
         #My Stuff
-        self.max_len = max_len
+        # self.label_embedding = label_embedding
+        # self.label_freeze = label_freeze
+        # self.d_a = d_a
+        # self.lstm_hid_dim = lstm_hid_dim
+
+        self._config["lstm_hid_dim"] = lstm_hid_dim
+        self._config["d_a"] = d_a
+        self._config["label_freeze"] = label_freeze
+        self._config["label_embedding"] = label_embedding
 
 
-        # Original
-        self.n_classes = len(classes)
-        self.embedding, self.tokenizer = get(representation, freeze=True)
-        self.embedding_dim = self.embedding(torch.LongTensor([[0]])).shape[-1]
-
-        if label_embed is not None:
-            self.label_embed = torch.nn.Embedding(label_embed.shape[0], label_embed.shape[1])
-            self.label_embed.from_pretrained(torch.FloatTensor(label_embed), freeze=label_freeze)
+        if self._config["label_embedding"] is not None:
+            self.label_embed = torch.nn.Embedding(
+                self._config["label_embedding"].shape[0],
+                self._config["label_embedding"].shape[1])
+            self.label_embed.from_pretrained(torch.FloatTensor(self._config["label_embedding"]), freeze=label_freeze)
         else:
-            self.label_embed = torch.nn.Embedding(self.n_classes, self.embedding_dim)
+            self.label_embed = torch.nn.Embedding(self.n_classes, self._config["lstm_hid_dim"])
 
-        self.lstm = torch.nn.LSTM(self.embedding_dim, hidden_size=lstm_hid_dim, num_layers=1,
-                                  batch_first=True, bidirectional=True)
+        if is_transformer(self.representation):
+            self.projection_input = torch.nn.Linear(self.embeddings_dim,
+                                                    self._config["lstm_hid_dim"] * 2)
+        else:
+            self.projection_input = torch.nn.LSTM(self.embeddings_dim,
+                                                  hidden_size=self._config["lstm_hid_dim"],
+                                                  num_layers=1,
+                                                  batch_first=True,
+                                                  bidirectional=True)
 
-        self.linear_first = torch.nn.Linear(lstm_hid_dim * 2, d_a)
-        self.linear_second = torch.nn.Linear(d_a, self.n_classes)
+        self.linear_first = torch.nn.Linear(self._config["lstm_hid_dim"] * 2, d_a)
+        self.linear_second = torch.nn.Linear(self._config["d_a"], self.n_classes)
 
-        self.weight1 = torch.nn.Linear(lstm_hid_dim * 2, 1)
-        self.weight2 = torch.nn.Linear(lstm_hid_dim * 2, 1)
+        self.weight1 = torch.nn.Linear(self._config["lstm_hid_dim"] * 2, 1)
+        self.weight2 = torch.nn.Linear(self._config["lstm_hid_dim"] * 2, 1)
 
-        self.output_layer = torch.nn.Linear(lstm_hid_dim * 2, self.n_classes)
+        self.output_layer = torch.nn.Linear(self._config["lstm_hid_dim"] * 2, self.n_classes)
         self.embedding_dropout = torch.nn.Dropout(p=0.5)
-        self.lstm_hid_dim = lstm_hid_dim
         self.build()
 
     def init_hidden(self, size):
-        return (torch.randn(2, size, self.lstm_hid_dim).to(self.device),
-                torch.randn(2, size, self.lstm_hid_dim).to(self.device))
+        """Deprecated"""
+        return (torch.randn(2, size, self._config["lstm_hid_dim"]).to(self.device),
+                torch.randn(2, size, self._config["lstm_hid_dim"]).to(self.device))
 
     def forward(self, x):
-        embeddings = self.embedding(x)
+        """
+        Forward pass function for transforming input tensor into output tensor.
+
+        :param x: Input tensor
+        :return: Output tensor
+        """
+        embeddings = self.embed_input(x) / self.embeddings_dim
         embeddings = self.embedding_dropout(embeddings)
         # step1 get LSTM outputs
         # hidden_state = self.init_hidden(x.shape[0])
-        outputs, hidden_state = self.lstm(embeddings)#, hidden_state)
+        outputs = self.projection_input(embeddings)
+        if not is_transformer(self._config["representation"]):
+            outputs = outputs[0]
         # step2 get self-attention
         selfatt = torch.tanh(self.linear_first(outputs))
         selfatt = self.linear_second(selfatt)
@@ -58,12 +87,12 @@ class LSANOriginal(TextClassificationAbstract):
         selfatt = selfatt.transpose(1, 2)
         self_att = torch.bmm(selfatt, outputs)
         # step3 get label-attention
-        h1 = outputs[:, :, :self.lstm_hid_dim]
-        h2 = outputs[:, :, self.lstm_hid_dim:]
+        h1 = outputs[:, :, :self._config["lstm_hid_dim"]]
+        h2 = outputs[:, :, self._config["lstm_hid_dim"]:]
 
         label = self.embedding_dropout(self.label_embed.weight.data)
-        m1 = torch.bmm(label.expand(x.shape[0], self.n_classes, self.lstm_hid_dim), h1.transpose(1, 2))
-        m2 = torch.bmm(label.expand(x.shape[0], self.n_classes, self.lstm_hid_dim), h2.transpose(1, 2))
+        m1 = torch.bmm(label.expand(x.shape[0], self.n_classes, self._config["lstm_hid_dim"]), h1.transpose(1, 2))
+        m2 = torch.bmm(label.expand(x.shape[0], self.n_classes, self._config["lstm_hid_dim"]), h2.transpose(1, 2))
         label_att = torch.cat((torch.bmm(m1, h1), torch.bmm(m2, h2)), 2)
         # label_att = F.normalize(label_att, p=2, dim=-1)
         # self_att = F.normalize(self_att, p=2, dim=-1) #all can
