@@ -1,6 +1,8 @@
 import torch
-from .abstracts import TextClassificationAbstract
+from .abstracts.abstracts import TextClassificationAbstract
 from ..representation import get
+from ..modules.module_KimCNN import KimCNNModule
+from ..representation import is_transformer
 ##############################################################################################
 ##############################################################################################
 #  Implementations
@@ -13,108 +15,70 @@ class XMLCNN(TextClassificationAbstract):
     Added support for Language Model as a feature.
     """
 
-    def __init__(self, classes, mode="transformer", representation="roberta", bottle_neck= None, kernel_sizes=[3, 4, 5, 6], filters=100,
-                 dropout=0.5, max_len=200,n_layers=4, **kwargs):
+    def __init__(self, mode="transformer",  bottle_neck= None, kernel_sizes=[3, 4, 5, 6], filters=100,
+                 dropout=0.5, **kwargs):
         """
-        Class constructor and intialization of every hyperparameters
-        :param classes:  A dictionary of the class label and the corresponding index
-        :param mode:  One of (trainable, untrainable, multichannel, transformer).
-        Transformer has to be used in combination with representation being a transformer model name (see: https://huggingface.co/transformers/pretrained_models.html).
-        In combination with the other three one of the glove embeddings can be used (glove50, glove100, glove200, glove300)
-        'trainable' will finetune the wordembedding used, whereas "untrainable" will freeze the embedding layer.
-        'multichannel' will combine two embedding layers, one for finetuning on the task, one frozen.
-        :param representation: The name of the representation to use. glove* or one of the hugginface transformers models.
+        Class constructor and initialization of every hyperparameter.
+
+        :param mode: One of (trainable, untrainable, multichannel, transformer).
+                    Transformer has to be used in combination with representation being a transformer model name (see: https://huggingface.co/transformers/pretrained_models.html).
+                    In combination with the other three one of the glove embeddings can be used (glove50, glove100, glove200, glove300)
+                    'trainable' will finetune the word embedding used, whereas "untrainable" will freeze the embedding layer.
+                    'multichannel' will combine two embedding layers, one for finetuning on the task, one frozen.
+        :param bottle_neck: Number of hidden units of the bottleneck layer
         :param kernel_sizes: Sizes of the kernel used for the convolution
         :param filters: Number of filters used in the convolution
-        :param dropout: Droupout rate
-        :param max_len: Maximum length input sequences. Longer sequences will be cut.
-        :param kwargs:
+        :param dropout: Dropout rate
+        :param kwargs: Optimizer and loss function keyword arguments, see `mlmc.models.abstracts.abstracts.TextClassificationAbstract`
         """
         super(XMLCNN, self).__init__(**kwargs)
 
-        self.classes = classes
-        self.n_classes = len(classes)
-        self.max_len = max_len
+        self._config["kernel_sizes"] = kernel_sizes
+        self._config["filters"] = filters
+        self._config["dropout"] = dropout
+        self._config["bottle_neck"] = bottle_neck if bottle_neck is not None else int(self.n_classes**0.5)
+        self._config["mode"] = mode
+
         self.modes = ("trainable", "untrainable", "multichannel", "transformer")
-        self.mode = mode
+        if is_transformer(self.representation):
+            print("Setting mode to transformer")
+            self._config["mode"] = "transformer"
+        assert self._config["mode"] in self.modes, "%s not in (%s, %s, %s, %s, %s)" % (self.mode, *self.modes)
+
         self.l = 1
-        self.kernel_sizes = kernel_sizes
-        self.filters = filters
-        self.representation = representation
-        self.dropout = dropout
-        self.n_layers = n_layers
-        self.bottle_neck = bottle_neck if bottle_neck is not None else int(len(classes)**0.5)
+        if self._config["mode"] =="multichannel":
+            self.l = 2
+            self.embedding_channel2, self.tokenizer_channel2 = get(model=self.representation, freeze=not self.finetune)
 
-        assert self.mode in self.modes, "%s not in (%s, %s, %s, %s)" % (self.mode, *self.modes)
-        self._init_input_representations()
 
-        self.convs = torch.nn.ModuleList(
-            [torch.nn.Conv1d(self.embeddings_dim, self.filters, k) for k in self.kernel_sizes])
-        self.projection = torch.nn.Linear(in_features=self.bottle_neck,
-                                          out_features=self.n_classes)
-        self.dropout_layer = torch.nn.Dropout(self.dropout)
-        self.bottle_neck_layer = torch.nn.Linear(in_features=self.l * len(self.kernel_sizes) * self.filters,
-                                                 out_features=self.bottle_neck)
+
+        # Layers
+        self.kimcnn_module = KimCNNModule(in_features=self.embeddings_dim,
+                                          kernel_sizes=self._config["kernel_sizes"],
+                                          filters=self._config["filters"],
+                                          dropout=self._config["dropout"])
+        self.dropout_layer = torch.nn.Dropout(self._config["dropout"])
+        self.bottle_neck_layer = torch.nn.Linear(
+            in_features=self.l * self.kimcnn_module.out_features,
+            out_features=self._config["bottle_neck"])
+        self.projection = torch.nn.Linear(
+            in_features=self._config["bottle_neck"],
+            out_features=self.n_classes)
+
         self.build()
 
-    def _init_input_representations(self):
-        if self.mode == "trainable":
-            self.embedding, self.tokenizer = get(model=self.representation, freeze=False)
-            self.embeddings_dim = self.embedding.weight.shape[-1]
-        elif self.mode == "untrainable":
-            self.embedding, self.tokenizer = get(model=self.representation, freeze=True)
-            self.embeddings_dim = self.embedding.weight.shape[-1]
-        elif self.mode == "multichannel":
-            self.l = 2
-            self.embedding, self.tokenizer = get(model=self.representation, freeze=True)
-            self.embeddings_dim = self.embedding.weight.shape[-1]
-
-            self.embedding_untrainable = torch.nn.Embedding(*self.embedding.weight.shape)
-            self.embedding_untrainable = self.embedding.from_pretrained(self.embedding.weight.clone(), freeze=False)
-            self.embedding = torch.nn.ModuleList([self.embedding, self.embedding_untrainable])
-
-        elif self.mode == "transformer":
-            try:
-                if self.n_layers == 1:
-                    self.embedding, self.tokenizer = get(model=self.representation)
-                    self.embeddings_dim = self.embedding(torch.tensor([[0]]))[0].shape[-1]
-                else:
-                    self.embedding, self.tokenizer = get(model=self.representation, output_hidden_states=True)
-                    self.embeddings_dim = torch.cat(self.embedding(self.embedding.dummy_inputs["input_ids"])[2][self.n_layers:], -1).shape[-1]
-            except TypeError:
-                Warning("If your using a model that does not support returning hiddenstates, set n_layers=1")
-                import sys
-                sys.exit()
-
     def forward(self, x):
+        """
+        Forward pass function for transforming input tensor into output tensor.
 
-        if self.mode == "trainable":
-            embedded = self.embedding(x).permute(0, 2, 1)
-            embedded = self.dropout_layer(embedded)
-            c = [torch.nn.functional.relu(conv(embedded).permute(0, 2, 1).max(1)[0]) for conv in self.convs]
-        elif self.mode == "untrainable":
-            with torch.no_grad():
-                embedded = self.embedding(x).permute(0, 2, 1)
-            embedded = self.dropout_layer(embedded)
-            c = [torch.nn.functional.relu(conv(embedded).permute(0, 2, 1).max(1)[0]) for conv in self.convs]
-        elif self.mode == "multichannel":
-            embedded_1 = self.embedding[0](x).permute(0, 2, 1)
-            with torch.no_grad():
-                embedded_2 = self.embedding[1](x).permute(0, 2, 1)
-            embedded_1 = self.dropout_layer(embedded_1)
-            embedded_2 = self.dropout_layer(embedded_2)
-            c = [torch.nn.functional.relu(conv(embedded_1).permute(0, 2, 1).max(1)[0]) for conv in self.convs] + \
-                [torch.nn.functional.relu(conv(embedded_2).permute(0, 2, 1).max(1)[0]) for conv in self.convs]
-        elif self.mode == "transformer":
-            if self.n_layers == 1:
-                with torch.no_grad():
-                    embedded = self.embedding(x)[0].permute(0, 2, 1)
-            else:
-                with torch.no_grad():
-                    embedded = torch.cat(self.embedding(x)[2][self.n_layers:], -1).permute(0, 2, 1)
-            c = [torch.nn.functional.relu(conv(embedded).permute(0, 2, 1).max(1)[0]) for conv in self.convs]
-
-        c = torch.cat(c, 1)
+        :param x: Input tensor
+        :return: Output tensor
+        """
+        e = self.embed_input(x)
+        c = self.kimcnn_module(e.permute(0, 2, 1))
+        if self._config["mode"] == "multichannel":
+            e2 = self.embedding_channel2(x)
+            c = torch.cat([c, self.kimcnn_module(e2.permute(0, 2, 1))], -1)
         c = torch.relu(self.bottle_neck_layer(c))
         output = self.projection(self.dropout_layer(c))
         return output
