@@ -14,27 +14,25 @@ def assertion_function(model_type, **kwargs):
     )
 
 
-    model = model_type(classes, **kwargs,
+    model = model_type(classes=classes, **kwargs,
                                optimizer_params={"lr": 5})
     history = model.fit(train=data,  epochs=15, batch_size=32, patience=2, tolerance=0.1)
 
     assert len(history) < 15, "The probabilistic test of early stopping failed. Try to re-run. If the error persists, it's bad."
 
-    eval = model.evaluate(data, return_report=True, return_roc=True)
+    loss, eval = model.evaluate(data)
     assert isinstance(eval, dict), "Return value of evaluate function failed."
-    assert isinstance(eval["report"], dict), "Evaluation Report is not a dict"
-    assert len(eval["report"]) == len(classes)+4, "Evaluation Report is not a dict"
 
-    model = model_type(classes, **kwargs, optimizer_params={"lr": 0.001})
+    model = model_type(classes=classes, **kwargs, optimizer_params={"lr": 0.001})
     history = model.fit(train=data, epochs=5, batch_size=3)
     assert len(history["train"]["loss"]) == 5, "Number of Epochs not reached"
 
 def test_KimCNN():
-    assertion_function(model_type=mlmc.models.KimCNN, mode="untrainable", representation="test")
+    assertion_function(model_type=mlmc.models.KimCNN, mode="untrainable", representation="test", filters=10, kernel_sizes=[3,4])
 
 def test_XMLCNN():
-    assertion_function(model_type=mlmc.models.XMLCNN, representation="test", mode="untrainable")
+    assertion_function(model_type=mlmc.models.XMLCNN, representation="test", mode="untrainable", filters=10, kernel_sizes=[3,4])
 
 def test_LSAN_transformer():
-    assertion_function(model_type=mlmc.models.LSANOriginalTransformerNoClasses, representation="roberta", n_layers=1)
+    assertion_function(model_type=mlmc.models.LSANNC, representation="test", label_model="test",n_layers=1, hidden_representations=10, d_a=10)
 
